Make some changes leading up to local/non-local. Seeems mostly functional now.
Changed analyzeDamage and checkLevelSetGradient to run on references to the vector, not the ambiguous choice of input vs global (separated the names of the internal variables).

Created a nucleated global variable. Can’t make more segments until nucleation has occurred.

Ensure that phi > phi_1. This can be an issue on the border of TLS and non-TLS as phi only gets fed once (half what it should be).

Nucleation peak values will be the maximum of current and proposed. No need to decrease phi by nucleation (only occurs in local/non-local model).

In analyzeDamage, removed non-TLS nodes from TLS segments. This is determined by seeing if the phi is already greater than the proposed phinew - if phi is greater than the value than any segment’s level-set would propose, it obviously doesn’t belong to a segment -> nonTLS.