Rewrite calculateYbar to make it cleaner. Basically, you need to find the beginning of the integrable part of the element and the end of the integrable part of the element. This is a function of the endpoints of the element, clearly, but peaks & minimums, and the segment zero location can further restrict the domain. Then, interpolate to integrate the residual and tangent over this found domain.
Also, disable standalone calculateYbar, since it makes a 'seemingly' inaccurate answer. In time steps where damage grows, the solution to Ybar-Yc = 0 is found, so Ybar/Yc should be 1, BUT after the segment integration boundaries are updated (peaks/mins updates as well as zero), a re-calculation of Ybar produces a different answer. By disabling this, I endure that the Ybar/Yc that should be 1 is the one reported.