More changes associated with implementing new softening model. Most notably implementing the LIN softening model for the local damage update and changing the H, dH, d2H functions to take phi as an input argument rather than d; the LIN model is much more easily expressed with phi.  LIN softening damage update is a 7th order polynomial, so a Newton-Raphson iteration is implemented.
In order to have H,dH,d2H in terms of phi, it becomes convenient to record phi (in addition to d) at the quadrature points. This may make the Y calculation (in calculateEnergies) a tad bit more accurate, but I don't think it matters.

Also, when d = 1, dH can be infinite or indeterminate. This isn't a problem for the non-local, since we'd multiply by d' which is zero, but for the local model, this screws up the Y calculation in calculateEnergies, so I threw in a check to clean that up.